// config/env.enum.ts
export enum EnvKeysEnum {
  NODE_ENV = 'NODE_ENV',
  SERVER_HTTP_PORT = 'SERVER_HTTP_PORT',
  SERVER_HTTP_HOST = 'SERVER_HTTP_HOST',
  DB_DIALECT = 'DB_DIALECT',
  DB_PORT = 'DB_PORT',
  DB_DATABASE = 'DB_DATABASE',
  DB_HOST = 'DB_HOST',
  DB_USERNAME = 'DB_USERNAME',
  DB_PASSWORD = 'DB_PASSWORD',
  DB_AUTO_LOAD_MODELS = 'DB_AUTO_LOAD_MODELS',
  DB_SYNC = 'DB_SYNC',
  DB_FORCE = 'DB_FORCE',
  DB_POOL_MIN = 'DB_POOL_MIN',
  DB_POOL_MAX = 'DB_POOL_MAX',
  DB_LOGGING = 'DB_LOGGING',
  DB_UNDERSCORED = 'DB_UNDERSCORED',
  KAFKA_CONSUMER_BROKERS = 'KAFKA_CONSUMER_BROKERS',
  KAFKA_PRODUCER_BROKERS = 'KAFKA_PRODUCER_BROKERS',
  KAFKA_CLIENT_ID = 'KAFKA_CLIENT_ID',
  KAFKA_GROUP_ID = 'KAFKA_GROUP_ID',
  KAFKA_ALLOW_AUTO_TOPIC_CREATION = 'KAFKA_ALLOW_AUTO_TOPIC_CREATION',
  KAFKA_AUTO_COMMIT = 'KAFKA_AUTO_COMMIT',
  KAFKA_FROM_BEGINING = 'KAFKA_FROM_BEGINING',
  KAFKA_TOPIC = 'KAFKA_TOPIC',
  GOOGLE_KEY = 'GOOGLE_KEY',
  BLOB_URL = 'BLOB_URL',
  BLOB_SAS_TOKEN = 'BLOB_SAS_TOKEN',
  BLOB_CONNECTION_STRING = 'BLOB_CONNECTION_STRING',
  BLOB_CONTAINER_NAME = 'BLOB_CONTAINER_NAME',
  CDN_LINK = 'CDN_LINK',
  CDN_AUTHORIZATION = 'CDN_AUTHORIZATION',
  DOMAIN_URL = 'DOMAIN_URL',
  IS_SWAGGER_ENABLED = 'IS_SWAGGER_ENABLED',
  OTEL_EXPORTER_OTLP_ENDPOINT = 'OTEL_EXPORTER_OTLP_ENDPOINT',
  SERVICE_NAME = 'SERVICE_NAME',
  KAFKA_AUDIT_LOG_TOPIC = 'KAFKA_AUDIT_LOG_TOPIC',
  AUDIT_LOG_NODE_HERO_IMAGE = 'AUDIT_LOG_NODE_HERO_IMAGE',
  GEMINI_API_KEY = 'GEMINI_API_KEY',
}